backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gut-iqapi"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "pydantic-settings",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.24",
    "pytest-mock",
    "freezegun",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]